0. Rebase on most recent master, 1. Remove err buffer shit, 2. complete the depth filter node so that there is an alternate scan, 3. new subscription and cb on simple_PID, 4. 'scan_occluded' and 'seek_center' now calc on the alternate scan
